Memory updates

+++ src/process.cpp
@@ -21,7 +21,7 @@ int Process::Pid() const {
 
 // Return this process's CPU utilization
 float Process::CpuUtilization() { 
-    long hertz = sysconf(_SC_CLK_TCK);
+    float hertz = sysconf(_SC_CLK_TCK);
     return (LinuxParser::ActiveJiffies(this->Pid())/hertz) / this->UpTime(); 
 }
